Remove custom db (#61)
* First batch of code deletion

* Remove more dead code

* swiftformat

* Update CHANGELOG.md

* Get rid of protocol indirection

* Remove direct GRDB dependency
