+++ GrailDiary/DocumentListViewController.swift
@@ -5,7 +5,6 @@ import CodableCSV
 import Combine
 import CoreServices
 import CoreSpotlight
-import GRDB
 import Logging
 import SafariServices
 import SnapKit
