+++ GrailDiary/DocumentTableController.swift
@@ -2,7 +2,6 @@
 
 import BookKit
 import Combine
-import GRDB
 import Logging
 import TextMarkupKit
 import UIKit
