+++ GrailDiary/QuotesViewController.swift
@@ -1,7 +1,6 @@
 // Copyright (c) 2018-2021  Brian Dewey. Covered by the Apache 2.0 license.
 
 import Combine
-import GRDB
 import Logging
 import SnapKit
 import TextMarkupKit
